+++ src/components/index.tsx
@@ -3,3 +3,5 @@ export * from './header/header';
 export * from './sidebar';
 export * from './achievement-card';
 export * from './badge-modal';
+export * from './productCard/productCard';
+export * from './searchBar/search';
