+++ tsconfig.json
@@ -25,6 +25,6 @@
       "@/cypress/*": ["./cypress/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/constants/index.tsx"],
   "exclude": ["node_modules"]
 }
